refactor(infra): rename lds stack file
